moved observation processing code to separate filea.
