+++ include/tudat/simulation/estimation_setup/observations.h
@@ -26,230 +26,13 @@
 #include "tudat/astro/observation_models/linkTypeDefs.h"
 #include "tudat/astro/observation_models/observableTypes.h"
 #include "tudat/simulation/estimation_setup/observationOutput.h"
+#include "tudat/simulation/estimation_setup/observationsProcessing.h"
 
 namespace tudat
 {
 
 namespace observation_models
 {
-
-enum ObservationFilterType
-{
-    residual_filtering,
-    absolute_value_filtering,
-    epochs_filtering,
-    time_bounds_filtering,
-    dependent_variable_filtering
-};
-
-struct ObservationFilterBase
-{
-public:
-
-    ObservationFilterBase( const ObservationFilterType filterType, const bool filterOut = true, const bool useOppositeCondition = false ) :
-            filterType_( filterType ), filterOut_( filterOut ), useOppositeCondition_( useOppositeCondition )
-    { }
-
-    virtual ~ObservationFilterBase( ){ }
-
-    ObservationFilterType getFilterType( ) const
-    {
-        return filterType_;
-    }
-
-    const bool filterOut( ) const
-    {
-        return filterOut_;
-    }
-
-    bool useOppositeCondition( ) const
-    {
-        return useOppositeCondition_;
-    }
-
-protected:
-
-    ObservationFilterType filterType_;
-    const bool filterOut_;
-    const bool useOppositeCondition_;
-};
-
-template< typename FilterValueType >
-bool checkFilterTypeConsistency( ObservationFilterType filterType )
-{
-    bool consistentType = true;
-    switch( filterType )
-    {
-        case residual_filtering:
-        case absolute_value_filtering:
-        {
-            if ( !( std::is_same< FilterValueType, double>::value ) ){ consistentType = false; }
-            break;
-        }
-        case epochs_filtering:
-        {
-            if ( !( std::is_same< FilterValueType, std::vector< double > >::value ) ){ consistentType = false; }
-            break;
-        }
-        case time_bounds_filtering:
-        {
-            if ( !( std::is_same< FilterValueType, std::pair< double, double > >::value ) ) { consistentType = false; }
-            break;
-        }
-        default:
-            break;
-    }
-    return consistentType;
-}
-
-
-template< typename FilterValueType >
-struct ObservationFilter : public ObservationFilterBase
-{
-public:
-    ObservationFilter( ObservationFilterType filterType, const FilterValueType& filterValue, const bool filterOut = true, const bool useOppositeCondition = false ) :
-            ObservationFilterBase( filterType, filterOut, useOppositeCondition ), filterValue_( filterValue )
-    {
-        if ( !checkFilterTypeConsistency< FilterValueType >( filterType ) )
-        {
-            throw std::runtime_error( "Error when creating observation filter of type " + std::to_string( filterType ) + " filter value type is inconsistent." );
-        }
-    }
-
-    virtual ~ObservationFilter( ){ }
-
-    FilterValueType getFilterValue( ) const
-    {
-        return filterValue_;
-    }
-
-protected:
-
-    FilterValueType filterValue_;
-};
-
-
-inline std::shared_ptr< ObservationFilterBase > observationFilter(
-        const ObservationFilterType filterType, const double filterValue, const bool filterOut = true, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationFilter< double > >( filterType, filterValue, filterOut, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationFilterBase > observationFilter(
-        const ObservationFilterType filterType, const std::vector< double > filterValue, const bool filterOut = true, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationFilter< std::vector< double > > >( filterType, filterValue, filterOut, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationFilterBase > observationFilter(
-        const ObservationFilterType filterType, const std::pair< double, double > filterValue, const bool filterOut = true, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationFilter< std::pair< double, double > > >( filterType, filterValue, filterOut, useOppositeCondition );
-}
-
-enum ObservationSetSplitterType
-{
-    time_tags_splitter,
-    time_interval_splitter,
-    time_span_splitter,
-    nb_observations_splitter,
-    dependent_variables_splitter
-};
-
-struct ObservationSetSplitterBase
-{
-public:
-    ObservationSetSplitterBase( const ObservationSetSplitterType splitterType, const int minNumberObservations = 0 ) :
-    splitterType_( splitterType ), minNumberObservations_( minNumberObservations ){ }
-
-    virtual ~ObservationSetSplitterBase( ){ }
-
-    ObservationSetSplitterType getSplitterType( ) const
-    {
-        return splitterType_;
-    }
-
-    int getMinNumberObservations( ) const
-    {
-        return minNumberObservations_;
-    }
-
-protected:
-    ObservationSetSplitterType splitterType_;
-    int minNumberObservations_;
-};
-
-template< typename SetSplitType >
-bool checkObsSetSplitTypeConsistency( ObservationSetSplitterType setSplitType )
-{
-    bool consistentType = true;
-    switch( setSplitType )
-    {
-        case time_tags_splitter:
-        {
-            if ( !( std::is_same< SetSplitType, std::vector< double > >::value ) ){ consistentType = false; }
-            break;
-        }
-        case time_interval_splitter:
-        case time_span_splitter:
-        {
-            if ( !( std::is_same< SetSplitType, double >::value ) ){ consistentType = false; }
-            break;
-        }
-        case nb_observations_splitter:
-        {
-            if ( !( std::is_same< SetSplitType, int >::value ) ) { consistentType = false; }
-            break;
-        }
-        default:
-            break;
-    }
-    return consistentType;
-}
-
-template< typename SplitterValueType >
-struct ObservationSetSplitter : public ObservationSetSplitterBase
-{
-public:
-    ObservationSetSplitter( ObservationSetSplitterType splitterType, const SplitterValueType& splitterValue, const int minNumberObservations ) :
-            ObservationSetSplitterBase( splitterType, minNumberObservations ), splitterValue_( splitterValue )
-    {
-        if ( !checkObsSetSplitTypeConsistency< SplitterValueType >( splitterType ) )
-        {
-            throw std::runtime_error( "Error when creating observation set splitter of type " + std::to_string( splitterType ) + " split value type is inconsistent." );
-        }
-    }
-
-    virtual ~ObservationSetSplitter( ){ }
-
-    SplitterValueType getSplitterValue( ) const
-    {
-        return splitterValue_;
-    }
-
-protected:
-
-    SplitterValueType splitterValue_;
-};
-
-inline std::shared_ptr< ObservationSetSplitterBase > observationSetSplitter(
-        const ObservationSetSplitterType splitterType, const std::vector< double > splitterValue, const int minNumberObservations = 0 )
-{
-    return std::make_shared< ObservationSetSplitter< std::vector< double > > >( splitterType, splitterValue, minNumberObservations );
-}
-
-inline std::shared_ptr< ObservationSetSplitterBase > observationSetSplitter(
-        const ObservationSetSplitterType splitterType, const double splitterValue, const int minNumberObservations = 0 )
-{
-    return std::make_shared< ObservationSetSplitter< double > >( splitterType, splitterValue, minNumberObservations );
-}
-
-inline std::shared_ptr< ObservationSetSplitterBase > observationSetSplitter(
-        const ObservationSetSplitterType splitterType, const int splitterValue, const int minNumberObservations = 0 )
-{
-    return std::make_shared< ObservationSetSplitter< int > >( splitterType, splitterValue, minNumberObservations );
-}
-
 
 template< typename ObservationScalarType = double, typename TimeType = double,
           typename std::enable_if< is_state_scalar_and_time_type< ObservationScalarType, TimeType >::value, int >::type = 0 >
@@ -929,7 +712,7 @@ private:
                 {
                     throw std::runtime_error( "Error when making SingleObservationSet, residuals size is incompatible after time ordering" );
                 }
-                std::map< TimeType, Eigen::Matrix< double, Eigen::Dynamic, 1 > > residualsMap;
+                std::map< TimeType, Eigen::Matrix< ObservationScalarType, Eigen::Dynamic, 1 > > residualsMap;
                 for( unsigned int i = 0; i < residuals_.size( ); i++ )
                 {
                     residualsMap[ observationTimes_.at( i ) ] = residuals_.at( i );
@@ -1245,269 +1028,6 @@ std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeT
 }
 
 
-enum ObservationParserType
-{
-    empty_parser,
-    observable_type_parser,
-    link_ends_parser,
-    link_end_id_parser,
-    time_bounds_parser,
-    ancillary_settings_parser,
-    multi_type_parser
-};
-
-struct ObservationCollectionParser
-{
-
-public:
-
-    ObservationCollectionParser( ) : parserType_( empty_parser ), useOppositeCondition_( false ){ }
-
-    ObservationCollectionParser( const ObservationParserType parserType, const bool useOppositeCondition = false ) :
-            parserType_( parserType ), useOppositeCondition_( useOppositeCondition ){ }
-
-    virtual ~ObservationCollectionParser( ){ }
-
-    ObservationParserType getObservationParserType( ) const
-    {
-        return parserType_;
-    }
-
-    bool useOppositeCondition( ) const
-    {
-        return useOppositeCondition_;
-    }
-
-protected:
-
-    const ObservationParserType parserType_;
-
-    const bool useOppositeCondition_;
-
-};
-
-struct ObservationCollectionObservableTypeParser : public ObservationCollectionParser
-{
-public:
-
-    ObservationCollectionObservableTypeParser( const ObservableType observableType, const bool useOppositeCondition = false ) :
-            ObservationCollectionParser( observable_type_parser, useOppositeCondition ), observableTypes_( std::vector< ObservableType >( { observableType } ) ){ }
-
-    ObservationCollectionObservableTypeParser( const std::vector< ObservableType > observableTypes, const bool useOppositeCondition = false ) :
-            ObservationCollectionParser( observable_type_parser, useOppositeCondition ), observableTypes_( observableTypes ){ }
-
-    virtual ~ObservationCollectionObservableTypeParser( ){ }
-
-    std::vector< ObservableType > getObservableTypes( ) const
-    {
-        return observableTypes_;
-    }
-
-protected:
-
-    const std::vector< ObservableType > observableTypes_;
-
-};
-
-struct ObservationCollectionLinkEndsParser : public ObservationCollectionParser
-{
-public:
-
-    ObservationCollectionLinkEndsParser( const LinkEnds linkEnds, const bool useOppositeCondition = false ) :
-            ObservationCollectionParser( link_ends_parser, useOppositeCondition ), linkEndsVector_( std::vector< LinkEnds >( { linkEnds } ) ){ }
-
-    ObservationCollectionLinkEndsParser( const std::vector< LinkEnds > linkEndsVector, const bool useOppositeCondition = false ) :
-            ObservationCollectionParser( link_ends_parser, useOppositeCondition ), linkEndsVector_( linkEndsVector ){ }
-
-    virtual ~ObservationCollectionLinkEndsParser( ){ }
-
-    std::vector< LinkEnds > getLinkEndsVector( ) const
-    {
-        return linkEndsVector_;
-    }
-
-protected:
-
-    const std::vector< LinkEnds > linkEndsVector_;
-
-};
-
-struct ObservationCollectionLinkEndIdParser : public ObservationCollectionParser
-{
-public:
-
-    ObservationCollectionLinkEndIdParser( const std::string linkEndsNames,
-                                          const bool isReferencePoint = false,
-                                          const bool useOppositeCondition = false ) :
-            ObservationCollectionParser( link_end_id_parser, useOppositeCondition ), linkEndsNames_( std::vector< std::string >( { linkEndsNames } ) ),
-            isReferencePoint_( isReferencePoint ){ }
-
-    ObservationCollectionLinkEndIdParser( const std::vector< std::string > linkEndsNames,
-                                          const bool isReferencePoint = false,
-                                          const bool useOppositeCondition = false ) :
-            ObservationCollectionParser( link_end_id_parser, useOppositeCondition ), linkEndsNames_( linkEndsNames ), isReferencePoint_( isReferencePoint ){ }
-
-    virtual ~ObservationCollectionLinkEndIdParser( ){ }
-
-    std::vector< std::string > getLinkEndNames( ) const
-    {
-        return linkEndsNames_;
-    }
-
-    bool isReferencePoint( ) const
-    {
-        return isReferencePoint_;
-    }
-
-protected:
-
-    const std::vector< std::string > linkEndsNames_;
-
-    const bool isReferencePoint_;
-
-};
-
-struct ObservationCollectionTimeBoundsParser : public ObservationCollectionParser
-{
-public:
-
-    ObservationCollectionTimeBoundsParser( const std::pair< double, double > timeBounds, const bool useOppositeCondition = false ) :
-            ObservationCollectionParser( time_bounds_parser, useOppositeCondition ), timeBoundsVector_( std::vector< std::pair< double, double > >( { timeBounds } ) ){ }
-
-    ObservationCollectionTimeBoundsParser( const std::vector< std::pair< double, double > > timeBoundsVector, const bool useOppositeCondition = false ) :
-            ObservationCollectionParser( time_bounds_parser, useOppositeCondition ), timeBoundsVector_( timeBoundsVector ){ }
-
-    virtual ~ObservationCollectionTimeBoundsParser( ){ }
-
-    std::vector< std::pair< double, double > > getTimeBoundsVector( ) const
-    {
-        return timeBoundsVector_;
-    }
-
-protected:
-
-    const std::vector< std::pair< double, double > > timeBoundsVector_;
-
-};
-
-struct ObservationCollectionAncillarySettingsParser : public ObservationCollectionParser
-{
-public:
-
-    ObservationCollectionAncillarySettingsParser( const std::shared_ptr< ObservationAncilliarySimulationSettings > ancillarySettings, const bool useOppositeCondition = false ) :
-            ObservationCollectionParser( ancillary_settings_parser, useOppositeCondition ),
-            ancillarySettings_( std::vector< std::shared_ptr< ObservationAncilliarySimulationSettings > >( { ancillarySettings } ) ){ }
-
-    ObservationCollectionAncillarySettingsParser( const std::vector< std::shared_ptr< ObservationAncilliarySimulationSettings > > ancillarySettings, const bool useOppositeCondition = false ) :
-            ObservationCollectionParser( ancillary_settings_parser, useOppositeCondition ), ancillarySettings_( ancillarySettings ){ }
-
-    virtual ~ObservationCollectionAncillarySettingsParser( ){ }
-
-    std::vector< std::shared_ptr< ObservationAncilliarySimulationSettings > > getAncillarySettings( ) const
-    {
-        return ancillarySettings_;
-    }
-
-protected:
-
-    const std::vector< std::shared_ptr< ObservationAncilliarySimulationSettings > > ancillarySettings_;
-
-};
-
-struct ObservationCollectionMultiTypeParser : public ObservationCollectionParser
-{
-public:
-
-    ObservationCollectionMultiTypeParser( const std::vector< std::shared_ptr< ObservationCollectionParser > >& observationParsers,
-                                          const bool combineConditions = false ) :
-            ObservationCollectionParser( multi_type_parser, false ), observationParsers_( observationParsers ), combineConditions_( combineConditions ){ }
-
-    virtual ~ObservationCollectionMultiTypeParser( ){ }
-
-    std::vector< std::shared_ptr< ObservationCollectionParser > > getObservationParsers_(  ) const
-    {
-        return observationParsers_;
-    }
-
-    bool areConditionsCombined( ) const
-    {
-        return combineConditions_;
-    }
-
-protected:
-
-    const std::vector< std::shared_ptr< ObservationCollectionParser > > observationParsers_;
-
-    const bool combineConditions_;
-
-};
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser( )
-{
-    return std::make_shared< ObservationCollectionParser >( );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser( const ObservableType observableType, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationCollectionObservableTypeParser >( observableType, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser( const std::vector< ObservableType >& observableTypes, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationCollectionObservableTypeParser >( observableTypes, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser( const LinkEnds linkEnds, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationCollectionLinkEndsParser >( linkEnds, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser( const std::vector< LinkEnds >& linkEndsVector, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationCollectionLinkEndsParser >( linkEndsVector, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser( const std::string bodyName, const bool isReferencePoint = false, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationCollectionLinkEndIdParser >( bodyName, isReferencePoint, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser( const std::vector< std::string >& bodyNames,
-                                                                         const bool isReferencePoint = false, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationCollectionLinkEndIdParser >( bodyNames, isReferencePoint, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser( const std::pair< double, double >& timeBounds, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationCollectionTimeBoundsParser >( timeBounds, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser( const std::vector< std::pair< double, double > >& timeBoundsVector, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationCollectionTimeBoundsParser >( timeBoundsVector, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser(
-        const std::shared_ptr< ObservationAncilliarySimulationSettings > ancillarySettings, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationCollectionAncillarySettingsParser >( ancillarySettings, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser(
-        const std::vector< std::shared_ptr< ObservationAncilliarySimulationSettings > >& ancillarySettings, const bool useOppositeCondition = false )
-{
-    return std::make_shared< ObservationCollectionAncillarySettingsParser >( ancillarySettings, useOppositeCondition );
-}
-
-inline std::shared_ptr< ObservationCollectionParser > observationParser( const std::vector< std::shared_ptr< ObservationCollectionParser > >& observationParsers,
-                                                                         const bool combineConditions = false )
-{
-    return std::make_shared< ObservationCollectionMultiTypeParser >( observationParsers, combineConditions );
-}
-
-
-
 template< typename ObservationScalarType = double, typename TimeType = double,
     typename std::enable_if< is_state_scalar_and_time_type< ObservationScalarType, TimeType >::value, int >::type = 0 >
 std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeType > >
@@ -1683,7 +1203,7 @@ public:
         return totalObservableSize_;
     }
 
-    SortedObservationSets getObservations( )
+    SortedObservationSets getObservationsSets( )
     {
         return observationSetList_;
     }
@@ -1992,10 +1512,10 @@ public:
         return concatenatedWeights;
     }
 
-    std::vector< Eigen::VectorXd > getResiduals(
+    std::vector< Eigen::Matrix< ObservationScalarType, Eigen::Dynamic, 1 > > getResiduals(
             std::shared_ptr< ObservationCollectionParser > observationParser = std::make_shared< ObservationCollectionParser >( ) ) const
     {
-        std::vector< Eigen::VectorXd > residuals;
+        std::vector< Eigen::Matrix< ObservationScalarType, Eigen::Dynamic, 1 > > residuals;
 
         std::map< ObservableType, std::map< LinkEnds, std::vector< unsigned int > > > observationSetsIndices = getSingleObservationSetsIndices( observationParser );
         for ( auto observableIt : observationSetsIndices )
@@ -2011,10 +1531,10 @@ public:
         return residuals;
     }
 
-    Eigen::VectorXd getConcatenatedResiduals(
+    Eigen::Matrix< ObservationScalarType, Eigen::Dynamic, 1 > getConcatenatedResiduals(
             std::shared_ptr< ObservationCollectionParser > observationParser = std::make_shared< ObservationCollectionParser >( ) ) const
     {
-        std::vector< Eigen::VectorXd > residuals = getResiduals( observationParser );
+        std::vector< Eigen::Matrix< ObservationScalarType, Eigen::Dynamic, 1 > > residuals = getResiduals( observationParser );
 
         unsigned int totalObsSize = 0;
         for ( auto residual : residuals )
@@ -2022,7 +1542,7 @@ public:
             totalObsSize += residual.size( );
         }
 
-        Eigen::VectorXd concatenatedResiduals = Eigen::VectorXd::Zero( totalObsSize );
+        Eigen::Matrix< ObservationScalarType, Eigen::Dynamic, 1 > concatenatedResiduals = Eigen::Matrix< ObservationScalarType, Eigen::Dynamic, 1 >::Zero( totalObsSize, 1 );
         unsigned int obsIndex = 0;
         for ( unsigned int k = 0 ; k < residuals.size( ) ; k++ )
         {
@@ -2199,7 +1719,7 @@ public:
         multiTypeParserList.push_back( observationParser( linkEndsDefinition.linkEnds_ ) );
         std::pair< Eigen::Matrix< ObservationScalarType, Eigen::Dynamic, 1 >, std::vector< TimeType > > observationsAndTimes =
                 getConcatenatedObservationsAndTimes( observationParser( multiTypeParserList, true ) );
-        if ( observationsAndTimes.empty( ) )
+        if ( observationsAndTimes.second.empty( ) )
         {
             throw std::runtime_error( " Error when getting single link observations and times, not observation set of type "
                                       + std::to_string( observableType ) + " for given link ends." );
@@ -2224,8 +1744,8 @@ public:
     }
 
     void setConstantWeight(
-            const std::shared_ptr< ObservationCollectionParser > observationParser,
-            const Eigen::VectorXd weight )
+            const Eigen::VectorXd weight,
+            const std::shared_ptr< ObservationCollectionParser > observationParser = std::make_shared< ObservationCollectionParser >( ) )
     {
         std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeType > > > singleObsSets = getSingleObservationSets( observationParser );
         if ( singleObsSets.empty( ) )
@@ -2257,8 +1777,8 @@ public:
     }
 
     void setTabulatedWeights(
-            const std::shared_ptr< ObservationCollectionParser > observationParser,
-            const Eigen::VectorXd tabulatedWeights )
+            const Eigen::VectorXd tabulatedWeights,
+            const std::shared_ptr< ObservationCollectionParser > observationParser = std::make_shared< ObservationCollectionParser >( ) )
     {
         std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeType > > > singleObsSets = getSingleObservationSets( observationParser );
         if ( singleObsSets.empty( ) )
@@ -2304,7 +1824,7 @@ public:
     {
         for ( auto parserIt : weightsPerObservationParser )
         {
-            setTabulatedWeights( parserIt.first, parserIt.second );
+            setTabulatedWeights( parserIt.second, parserIt.first );
         }
     }
 
@@ -3064,7 +2584,7 @@ std::shared_ptr< ObservationCollection< ObservationScalarType, TimeType > > spli
     const int minimumNumberOfObservations )
 {
     typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets originalObservationSets =
-        originalObservationCollection->getObservations( );
+        originalObservationCollection->getObservationsSets( );
     typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets splitObservationSets;
     for( auto observationIt : originalObservationSets )
     {
@@ -3116,10 +2636,10 @@ std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeT
         const ObservableType observableType )
 {
     std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeType > > > observationList;
-    if( observationCollection->getObservations( ).count( observableType ) != 0 )
+    if( observationCollection->getObservationsSets( ).count( observableType ) != 0 )
     {
         std::map< LinkEnds, std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeType > > > >
-                observationsOfGivenType = observationCollection->getObservations( ).at( observableType );
+                observationsOfGivenType = observationCollection->getObservationsSets( ).at( observableType );
 
         for( auto linkEndIterator : observationsOfGivenType )
         {
@@ -3142,12 +2662,12 @@ std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeT
         const LinkEnds& linkEnds )
 {
     std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeType > > > observationList;
-    if( observationCollection->getObservations( ).count( observableType ) != 0 )
+    if( observationCollection->getObservationsSets( ).count( observableType ) != 0 )
     {
-        if( observationCollection->getObservations( ).at( observableType ).count( linkEnds ) != 0 )
+        if( observationCollection->getObservationsSets( ).at( observableType ).count( linkEnds ) != 0 )
         {
             std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeType > > > fullObservationList =
-                    observationCollection->getObservations( ).at( observableType ).at( linkEnds );
+                    observationCollection->getObservationsSets( ).at( observableType ).at( linkEnds );
             std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeType > > > partialObservationList =
                     getObservationListWithDependentVariables( fullObservationList, dependentVariableToRetrieve );
             observationList.insert( observationList.end( ), partialObservationList.begin( ), partialObservationList.end( ) );
@@ -3265,8 +2785,8 @@ std::shared_ptr< ObservationCollection< ObservationScalarType, TimeType > >  cre
     const std::shared_ptr< ObservationCollection< ObservationScalarType, TimeType > > computedData )
 {
 //    std::map< ObservableType, std::map< LinkEnds, std::vector< std::shared_ptr< SingleObservationSet< ObservationScalarType, TimeType > > > > >
-    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets observedObservationSets = observedData->getObservations( );
-    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets computedObservationSets = computedData->getObservations( );
+    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets observedObservationSets = observedData->getObservationsSets( );
+    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets computedObservationSets = computedData->getObservationsSets( );
     typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets residualObservationSets;
 
     for( auto observationIt : observedObservationSets )
@@ -3297,7 +2817,7 @@ std::shared_ptr< ObservationCollection< ObservationScalarType, TimeType > >  cre
 //    }
 //
 //    // Retrieve observations to filter
-//    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets observationSetsToFilter = dataToFiler->getObservations( );
+//    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets observationSetsToFilter = dataToFiler->getObservationsSets( );
 //
 //    // Create data structure with filtered results
 //    std::map< observation_models::ObservableType, std::vector< std::pair< LinkEnds, std::vector< std::vector< int > > > > > filterEntries;
@@ -3423,9 +2943,9 @@ std::shared_ptr< ObservationCollection< ObservationScalarType, TimeType > >  cre
 //    const std::shared_ptr< ObservationCollection< ObservationScalarType, TimeType > > residualData,
 //    const std::map< ObservableType, double > residualCutoffValuePerObservable )
 //{
-//    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets observedObservationSets = observedData->getObservations( );
+//    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets observedObservationSets = observedData->getObservationsSets( );
 //    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets filteredObservedObservationSets;
-//    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets residualObservationSets = residualData->getObservations( );
+//    typename ObservationCollection< ObservationScalarType, TimeType >::SortedObservationSets residualObservationSets = residualData->getObservationsSets( );
 //
 //    for( auto observationIt : residualObservationSets )
 //    {
