+++ tests/src/astro/observation_models/CMakeLists.txt
@@ -66,7 +66,7 @@ TUDAT_ADD_TEST_CASE(ObservationRoundingErrors PRIVATE_LINKS ${Tudat_ESTIMATION_L
 
 TUDAT_ADD_TEST_CASE(ObservationDependentVariables PRIVATE_LINKS ${Tudat_ESTIMATION_LIBRARIES})
 
-TUDAT_ADD_TEST_CASE(DsnOdfGrail PRIVATE_LINKS ${Tudat_ESTIMATION_LIBRARIES})
+#TUDAT_ADD_TEST_CASE(DsnOdfGrail PRIVATE_LINKS ${Tudat_ESTIMATION_LIBRARIES})
 
 #TUDAT_ADD_TEST_CASE(DsnOdfGrailStatistics PRIVATE_LINKS ${Tudat_ESTIMATION_LIBRARIES})
