+++ tests/src/astro/orbit_determination/CMakeLists.txt
@@ -131,6 +131,11 @@ TUDAT_ADD_TEST_CASE(NonSequentialStateEstimation
          ${Tudat_ESTIMATION_LIBRARIES}
  )
 
+ TUDAT_ADD_TEST_CASE(ObservationsProcessing
+         PRIVATE_LINKS
+         ${Tudat_ESTIMATION_LIBRARIES}
+ )
+
 if( TUDAT_BUILD_WITH_EXTENDED_PRECISION_PROPAGATION_TOOLS )
